pyproject.toml
[tool.poetry]
name = "repl_python3_Flask"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"
absl-py = "^1.0.0"
python-twitter-v2 = "^0.7.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
